+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-## Improvements 
-- Fixed unwanted navigation bar showing when using `.groupChannelView{ }` view builder inside a `TabView`
+## New Interfaces
+New view adaptor interfaces for `GroupChannelType.ListItem` have been added.
+- `public func quotedFileMessageView<Content: View>(@ViewBuilder content:)`
+  - An interface that customizes a quoted file message view.  
+- `public func quotedMultipleFilesMessageView<Content: View>(@ViewBuilder content:)`
+  - An interface that customizes a quoted multiple files message view.